Cmp return int64_t.

+++ src/cmap-ext.h
@@ -52,7 +52,7 @@ CMAP_MAP * cmap_fn_require_definitions(CMAP_FN * fn, CMAP_PROC_CTX * proc_ctx);
 
 CMAP_MAP * cmap_copy_map(CMAP_MAP * dst, CMAP_MAP * src);
 
-int cmap_cmp(CMAP_MAP * map_l, CMAP_MAP * map_r);
+int64_t cmap_cmp(CMAP_MAP * map_l, CMAP_MAP * map_r);
 
 CMAP_MAP * cmap_new(CMAP_FN * prototype, CMAP_PROC_CTX * proc_ctx, ...);
 CMAP_MAP * cmap_lnew(CMAP_FN * prototype, CMAP_PROC_CTX * proc_ctx,
