+++ src/core/cmp/cmap-cmp.h
@@ -1,11 +1,12 @@
 #ifndef __CMAP_CMP_H__
 #define __CMAP_CMP_H__
 
+#include <stdint.h>
 #include "cmap-map-type.h"
 
 typedef struct
 {
-  int (*cmp)(CMAP_MAP * map_l, CMAP_MAP * map_r);
+  int64_t (*cmp)(CMAP_MAP * map_l, CMAP_MAP * map_r);
 } CMAP_CMP_PUBLIC;
 
 extern const CMAP_CMP_PUBLIC cmap_cmp_public;
